abstract common methods for DB queries
